requirements.txt
Django
django-ninja
psycopg[binary]
postgis

django-allauth
django-defender
django-csp

django-storages
Pillow
django-imagekit

django-filter
django-environ
django-money
django-health-check

celery
redis
django-celery-beat

django-ninja-extra
django-ninja-jwt
django-cors-headers
orjson
django-redis

django-debug-toolbar
black
isort
flake8
pre-commit

pytest-django
factory-boy
model-bakery
